added readme

+++ ft_printf/ft_putaddrs_fd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putaddrs_fd.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pamallet <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: paul_mallet <paul_mallet@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 17:00:11 by pamallet          #+#    #+#             */
-/*   Updated: 2024/11/07 11:23:56 by pamallet         ###   ########.fr       */
+/*   Updated: 2026/01/13 22:18:24 by paul_mallet      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putaddrs_fd(unsigned long n, char *shex, int fd)
 {
-	if (n >= 0 && n <= 15)
+	if ((int)n >= 0 && n <= 15)
 		write(fd, &shex[n % 16], 1);
 	else
 	{
