+++ ft_printf/ft_putunbr_fd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putunbr_fd.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pamallet <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: paul_mallet <paul_mallet@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 10:31:33 by pamallet          #+#    #+#             */
-/*   Updated: 2024/11/07 11:40:46 by pamallet         ###   ########.fr       */
+/*   Updated: 2026/01/13 22:17:51 by paul_mallet      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putunbr_fd(unsigned int n, int fd)
 {
-	if (n >= 0 && n <= 9)
+	if ((int)n >= 0 && n <= 9)
 	{
 		n += 48;
 		write(fd, &n, 1);
